implemented separate click-event function
