+++ src/react-app-env.d.ts
@@ -1,6 +1,1 @@
 /// <reference types="react-scripts" />
-export interface Tab {
-  id: string,
-  title: string,
-  content: string,
-}
