+++ src/types/Tab.ts
@@ -1,0 +1,5 @@
+export interface Tab {
+  id: string,
+  title: string,
+  content: string,
+}
